test(helpers): re-organize tests to reflect code structure
Refs #2717
